docs: icon buttons
