+++ docs/src/components/ReactLib/Icons/PandoIcons.astro
@@ -1,0 +1,5 @@
+---
+import ReactPandoIcons from './PandoIcons'
+---
+
+<ReactPandoIcons client:only />
