+++ docs/src/content/docs/reference/icons.mdx
@@ -2,6 +2,6 @@
 title: Icons
 ---
 
-import PandoIcons from '@/src/components/ReactLib/Icons/ReactIcons.tsx'
+import PandoIcons from '@/src/components/ReactLib/Icons/PandoIcons.astro'
 
 <PandoIcons />
